Uses the dialog title theme
